requirements.mac.py312.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-model-utils>=4.5
django-notifications-hq>=1.8
django-ratelimit>=4.1
django-storages>=1.14
jsonfield>=3.1
swapper>=1.3

PyMySQL>=1.1.0

celery>=5.3
redis>=5.0

boto3>=1.34
botocore>=1.34
s3transfer>=0.10

python-dateutil>=2.9
requests>=2.31
urllib3>=2.2
pytz>=2024.1
sqlparse>=0.5

python-memcached>=1.62
